Renames lib.py to scan_lib.py
