+++ ccp-openshift-scan/scanning/scan_lib.py
